+++ src/main/java/by/jackraidenph/dragonsurvival/handlers/BlockInit.java
@@ -3,6 +3,7 @@ package by.jackraidenph.dragonsurvival.handlers;
 import by.jackraidenph.dragonsurvival.DragonSurvivalMod;
 import by.jackraidenph.dragonsurvival.blocks.DragonAltarBlock;
 import by.jackraidenph.dragonsurvival.blocks.DragonGateBlock;
+import by.jackraidenph.dragonsurvival.blocks.DragonGateController;
 import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.material.Material;
@@ -20,7 +21,7 @@ import net.minecraftforge.registries.ObjectHolder;
 @ObjectHolder(DragonSurvivalMod.MODID)
 @Mod.EventBusSubscriber(modid = DragonSurvivalMod.MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class BlockInit {
-    static ItemGroup blocks=new ItemGroup("Dragon Survival - blocks") {
+    static ItemGroup blocks=new ItemGroup("dragon.survival.blocks") {
         @Override
         public ItemStack createIcon() {
             return new ItemStack(dragon_altar);
@@ -35,19 +36,24 @@ public class BlockInit {
                     .hardnessAndResistance(5.0f)
                     .lightValue(5)
                     .sound(SoundType.ANVIL));
+
     static final DragonGateBlock dragonGate=new DragonGateBlock(Block.Properties.create(Material.WOOD).hardnessAndResistance(5,30));
+    static final DragonGateController dragonGateController=new DragonGateController(Block.Properties.create(Material.WOOD).hardnessAndResistance(5,30));
 
     @SubscribeEvent
     public static void registerBlocks(final RegistryEvent.Register<Block> event) {
         event.getRegistry().register(dragon_altar);
         IForgeRegistry<Block> forgeRegistry= event.getRegistry();
         forgeRegistry.register(dragonGate.setRegistryName(DragonSurvivalMod.MODID,"dragon_gate"));
+        forgeRegistry.register(dragonGateController.setRegistryName(DragonSurvivalMod.MODID,"dragon_gate_controller"));
+
     }
 
     @SubscribeEvent
     public static void registerBlockItems(final RegistryEvent.Register<Item> event) {
         event.getRegistry().register(new BlockItem(dragon_altar, new Item.Properties().group(ItemGroup.DECORATIONS)).setRegistryName("dragon_altar"));
         IForgeRegistry<Item> forgeRegistry= event.getRegistry();
-        forgeRegistry.register(new BlockItem(dragon_altar,new Item.Properties().group(blocks)).setRegistryName(dragonGate.getRegistryName()));
+        forgeRegistry.register(new BlockItem(dragonGate,new Item.Properties().group(blocks)).setRegistryName(dragonGate.getRegistryName()));
+        forgeRegistry.register(new BlockItem(dragonGateController,new Item.Properties().group(blocks)).setRegistryName(dragonGateController.getRegistryName()));
     }
 }
