+++ src/main/resources/assets/dragonsurvival/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "block.dragon_survival.dragon_gate": "Dragon Gate",
+  "block.dragon_survival.dragon_gate_controller": "Dragon Gate Controller"
+}
